requirements.txt
CurrencyConverter>=0.16
python-binance>=1.0.7
pycoingecko>=2.0.0
python-dotenv>=0.17.0
pytz>=2021.1
